+++ kt/plugins/godot-gradle-plugin/src/main/kotlin/godot/gradle/util/mapOfNonNullValuesOf.kt
@@ -1,0 +1,10 @@
+package godot.gradle.util
+
+fun mapOfNonNullValuesOf(vararg entries: Pair<String, String?>): Map<String, String> =
+    mutableMapOf<String, String>().apply {
+        for ((key, value) in entries) {
+            if (value != null) {
+                put(key, value)
+            }
+        }
+    }
